requirements.txt
appdirs==1.4.4
asgiref==3.7.2
attrs==23.1.0
axblack==20220330
click==8.1.6
Django==4.2.4
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mypy-extensions==1.0.0
pathspec==0.11.2
Pillow==10.0.0
PyJWT==2.8.0
pytz==2023.3
regex==2023.8.8
sqlparse==0.4.4
toml==0.10.2
typed-ast==1.5.5
typing_extensions==4.7.1
